requirements.txt
asgiref==3.2.3
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
numpy==1.18.1
opencv-python==4.2.0.32
pytz==2019.3
sqlparse==0.3.1
